Fixed ABC norm calculation

+++ src/RcppExports.cpp
@@ -62,7 +62,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // sample_ABC_rdirdirgamma_cpp
-RcppGSL::Matrix sample_ABC_rdirdirgamma_cpp(const unsigned int& n_sample, const unsigned int& m_sample, const double& alpha_0, const double& beta_0, const Rcpp::NumericVector& nu_0, const Rcpp::NumericMatrix& mtx_obs, const unsigned int& reps, const unsigned int& p_norm, const unsigned int seed);
+RcppGSL::Matrix sample_ABC_rdirdirgamma_cpp(const unsigned int& n_sample, const unsigned int& m_sample, const double& alpha_0, const double& beta_0, const Rcpp::NumericVector& nu_0, const Rcpp::NumericMatrix& mtx_obs, const unsigned int& reps, const double& p_norm, const unsigned int seed);
 RcppExport SEXP _rdirdirgamma_sample_ABC_rdirdirgamma_cpp(SEXP n_sampleSEXP, SEXP m_sampleSEXP, SEXP alpha_0SEXP, SEXP beta_0SEXP, SEXP nu_0SEXP, SEXP mtx_obsSEXP, SEXP repsSEXP, SEXP p_normSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -74,14 +74,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const Rcpp::NumericVector& >::type nu_0(nu_0SEXP);
     Rcpp::traits::input_parameter< const Rcpp::NumericMatrix& >::type mtx_obs(mtx_obsSEXP);
     Rcpp::traits::input_parameter< const unsigned int& >::type reps(repsSEXP);
-    Rcpp::traits::input_parameter< const unsigned int& >::type p_norm(p_normSEXP);
+    Rcpp::traits::input_parameter< const double& >::type p_norm(p_normSEXP);
     Rcpp::traits::input_parameter< const unsigned int >::type seed(seedSEXP);
     rcpp_result_gen = Rcpp::wrap(sample_ABC_rdirdirgamma_cpp(n_sample, m_sample, alpha_0, beta_0, nu_0, mtx_obs, reps, p_norm, seed));
     return rcpp_result_gen;
 END_RCPP
 }
 // sample_ABC_rdirdirgamma_beta_cpp
-Rcpp::NumericMatrix sample_ABC_rdirdirgamma_beta_cpp(const unsigned int& n_sample, const unsigned int& m_sample, const double& alpha_0, const double& beta_0, const Rcpp::NumericVector& nu_0, const Rcpp::NumericMatrix& mtx_obs, const unsigned int& reps, const unsigned int& p_norm);
+Rcpp::NumericMatrix sample_ABC_rdirdirgamma_beta_cpp(const unsigned int& n_sample, const unsigned int& m_sample, const double& alpha_0, const double& beta_0, const Rcpp::NumericVector& nu_0, const Rcpp::NumericMatrix& mtx_obs, const unsigned int& reps, const double& p_norm);
 RcppExport SEXP _rdirdirgamma_sample_ABC_rdirdirgamma_beta_cpp(SEXP n_sampleSEXP, SEXP m_sampleSEXP, SEXP alpha_0SEXP, SEXP beta_0SEXP, SEXP nu_0SEXP, SEXP mtx_obsSEXP, SEXP repsSEXP, SEXP p_normSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -93,7 +93,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const Rcpp::NumericVector& >::type nu_0(nu_0SEXP);
     Rcpp::traits::input_parameter< const Rcpp::NumericMatrix& >::type mtx_obs(mtx_obsSEXP);
     Rcpp::traits::input_parameter< const unsigned int& >::type reps(repsSEXP);
-    Rcpp::traits::input_parameter< const unsigned int& >::type p_norm(p_normSEXP);
+    Rcpp::traits::input_parameter< const double& >::type p_norm(p_normSEXP);
     rcpp_result_gen = Rcpp::wrap(sample_ABC_rdirdirgamma_beta_cpp(n_sample, m_sample, alpha_0, beta_0, nu_0, mtx_obs, reps, p_norm));
     return rcpp_result_gen;
 END_RCPP
